requirements.txt
boto3==1.17.53
botocore==1.20.53
docopt==0.6.2
jmespath==0.10.0
librosa==0.9.1
python-dateutil==2.8.1
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
webvtt-py==0.4.6
